chore: 사용하지 않는 코드 정리(#8)

+++ src/main/java/org/example/bogo/domain/project/presentation/dto/request/PostAnswerRequest.java
@@ -6,7 +6,7 @@ public record PostAnswerRequest(Long projectId,
                                 List<Answer> answers) {
 
     public record Answer(
-            int qId,
+            Long qId,
             String a
     ) {}
 }
